requirements.txt
pykrx>=1.0.45
finance-datareader>=0.9.66
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
pyyaml>=6.0
schedule>=1.2.0
requests>=2.31.0
